pyproject.toml
[tool.poetry]
name = "tpaneldesktop"
version = "0.1.0"
description = ""
authors = ["Georgy <goxa1@vk.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.0.3"
django-cors-headers = "4.3.1"
django-jazzmin = "2.6.1"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
pillow = "10.2.0"
pyjwt = "2.8.0"
sqlparse = "0.4.4"
tzdata = "2024.1"
python-dotenv = "^1.1.0"
django-redis = "^5.4.0"
redis = "^5.0.1"
hiredis = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
